feat(mlog): 更新日志管理，重构日志记录逻辑，新增结构化日志支持，优化日志钩子功能
