+++ os/mlog/mlog_api.go
@@ -2,78 +2,76 @@ package mlog
 
 import "context"
 
-// Print prints `v` with newline using fmt.Sprintln.
-// The parameter `v` can be multiple variables.
-func Print(ctx context.Context, v ...interface{}) {
-	defaultLogger.Print(ctx, v...)
-}
-
-// Printf prints `v` with format `format` using fmt.Sprintf.
-// The parameter `v` can be multiple variables.
-func Printf(ctx context.Context, format string, v ...interface{}) {
-	defaultLogger.Printf(ctx, format, v...)
-}
-
-// Fatal prints the logging content with [FATA] header and newline, then exit the current process.
-func Fatal(ctx context.Context, v ...interface{}) {
-	defaultLogger.Fatal(ctx, v...)
-}
-
-// Fatalf prints the logging content with [FATA] header, custom format and newline, then exit the current process.
-func Fatalf(ctx context.Context, format string, v ...interface{}) {
-	defaultLogger.Fatalf(ctx, format, v...)
-}
-
-// Panic prints the logging content with [PANI] header and newline, then panics.
-func Panic(ctx context.Context, v ...interface{}) {
-	defaultLogger.Panic(ctx, v...)
-}
-
-// Panicf prints the logging content with [PANI] header, custom format and newline, then panics.
-func Panicf(ctx context.Context, format string, v ...interface{}) {
-	defaultLogger.Panicf(ctx, format, v...)
-}
-
-// Info prints the logging content with [INFO] header and newline.
-func Info(ctx context.Context, v ...interface{}) {
-	defaultLogger.Info(ctx, v...)
-}
-
-// Infof prints the logging content with [INFO] header, custom format and newline.
-func Infof(ctx context.Context, format string, v ...interface{}) {
-	defaultLogger.Infof(ctx, format, v...)
-}
-
-// Debug prints the logging content with [DEBU] header and newline.
-func Debug(ctx context.Context, v ...interface{}) {
-	defaultLogger.Debug(ctx, v...)
-}
-
 // Debugf prints the logging content with [DEBU] header, custom format and newline.
-func Debugf(ctx context.Context, format string, v ...interface{}) {
+func Debugf(ctx context.Context, format string, v ...any) {
 	defaultLogger.Debugf(ctx, format, v...)
 }
 
-// Warn prints the logging content with [WARN] header and newline.
-// It also prints caller stack info if stack feature is enabled.
-func Warn(ctx context.Context, v ...interface{}) {
-	defaultLogger.Warn(ctx, v...)
+// Debugw prints the logging content with [DEBU] header, custom format and newline.
+func Debugw(ctx context.Context, msg string, fields ...Field) {
+	defaultLogger.Debugw(ctx, msg, fields...)
+}
+
+// Infof prints the logging content with [INFO] header, custom format and newline.
+func Infof(ctx context.Context, format string, v ...any) {
+	defaultLogger.Infof(ctx, format, v...)
+}
+
+// Info prints the logging content with [INFO] header and newline.
+func Infow(ctx context.Context, msg string, fields ...Field) {
+	defaultLogger.Infow(ctx, msg, fields...)
 }
 
 // Warnf prints the logging content with [WARN] header, custom format and newline.
-// It also prints caller stack info if stack feature is enabled.
-func Warnf(ctx context.Context, format string, v ...interface{}) {
+func Warnf(ctx context.Context, format string, v ...any) {
 	defaultLogger.Warnf(ctx, format, v...)
 }
 
-// Error prints the logging content with [ERRO] header and newline.
-// It also prints caller stack info if stack feature is enabled.
-func Error(ctx context.Context, v ...interface{}) {
-	defaultLogger.Error(ctx, v...)
+// Warnw prints the logging content with [WARN] header, custom format and newline.
+func Warnw(ctx context.Context, msg string, fields ...Field) {
+	defaultLogger.Warnw(ctx, msg, fields...)
 }
 
 // Errorf prints the logging content with [ERRO] header, custom format and newline.
-// It also prints caller stack info if stack feature is enabled.
-func Errorf(ctx context.Context, format string, v ...interface{}) {
-	defaultLogger.Errorf(ctx, format, v...)
+func Errorf(ctx context.Context, err error, format string, v ...any) {
+	defaultLogger.Errorf(ctx, err, format, v...)
+}
+
+// Errorw prints the logging content with [ERRO] header, custom format and newline.
+func Errorw(ctx context.Context, err error, msg string, fields ...Field) {
+	defaultLogger.Errorw(ctx, err, msg, fields...)
+}
+
+// Fatalf prints the logging content with [FATA] header, custom format and newline, then exit the current process.
+func Fatalf(ctx context.Context, err error, format string, v ...any) {
+	defaultLogger.Fatalf(ctx, err, format, v...)
+}
+
+// Fatalw prints the logging content with [FATA] header, custom format and newline, then exit the current process.
+func Fatalw(ctx context.Context, err error, msg string, fields ...Field) {
+	defaultLogger.Fatalw(ctx, err, msg, fields...)
+}
+
+// Panicf prints the logging content with [PANI] header, custom format and newline, then panics.
+func Panicf(ctx context.Context, err error, format string, v ...any) {
+	defaultLogger.Panicf(ctx, err, format, v...)
+}
+
+func Panicw(ctx context.Context, err error, msg string, fields ...Field) {
+	defaultLogger.Panicw(ctx, err, msg, fields...)
+}
+
+// With returns a new logger with the added attributes.
+func With(fields ...Field) *Logger {
+	return defaultLogger.With(fields...)
+}
+
+// AddHook adds a hook to the logger.
+func AddHook(hook Hook) {
+	defaultLogger.AddHook(hook)
+}
+
+// Close closes the logger and its underlying resources.
+func Close() error {
+	return defaultLogger.Close()
 }
