--- os/mlog/mlog_logger_hook_adapter.go
@@ -1,48 +1,0 @@
-package mlog
-
-import (
-	"maps"
-
-	"github.com/sirupsen/logrus"
-)
-
-// logrusHook is an adapter that wraps a mlog.Hook to make it compatible with the logrus.Hook interface.
-// This allows mlog to use its own simplified Hook interface while still leveraging the underlying logrus hooking mechanism.
-type logrusHook struct {
-	hook Hook
-}
-
-// Levels implements the logrus.Hook interface.
-func (h *logrusHook) Levels() []logrus.Level {
-	levels := h.hook.Levels()
-	logrusLevels := make([]logrus.Level, len(levels))
-	for i, level := range levels {
-		logrusLevels[i] = logrus.Level(level)
-	}
-	return logrusLevels
-}
-
-// Fire implements the logrus.Hook interface.
-func (h *logrusHook) Fire(entry *logrus.Entry) error {
-	// create mlog.Entry
-	e := &Entry{
-		Level:   Level(entry.Level),
-		Message: entry.Message,
-		Data:    make(map[string]any),
-		Context: entry.Context,
-		raw:     entry,
-	}
-
-	// copy data
-	maps.Copy(e.Data, entry.Data)
-
-	// call user's hook
-	if err := h.hook.Fire(e); err != nil {
-		return err
-	}
-
-	// copy data back to logrus entry
-	maps.Copy(entry.Data, e.Data)
-
-	return nil
-}
