+++ os/mlog/z_mlog_benchmark_test.go
@@ -2,6 +2,7 @@ package mlog_test
 
 import (
 	"context"
+	"os"
 	"testing"
 
 	"github.com/graingo/maltose/os/mlog"
@@ -11,48 +12,40 @@ var (
 	// A static context to avoid context creation overhead in benchmarks.
 	ctx = context.Background()
 	// Pre-allocate fields to avoid allocation overhead inside the benchmark loop itself.
-	tenLogrusFields = mlog.Fields{
-		"key1":  "value1",
-		"key2":  123,
-		"key3":  true,
-		"key4":  123.456,
-		"key5":  "value5",
-		"key6":  "value6",
-		"key7":  "value7",
-		"key8":  "value8",
-		"key9":  "value9",
-		"key10": "value10",
+	tenMlogFields = []mlog.Field{
+		mlog.String("key1", "value1"),
+		mlog.Int("key2", 123),
+		mlog.Bool("key3", true),
+		mlog.Float64("key4", 123.456),
+		mlog.String("key5", "value5"),
+		mlog.String("key6", "value6"),
+		mlog.String("key7", "value7"),
+		mlog.String("key8", "value8"),
+		mlog.String("key9", "value9"),
+		mlog.String("key10", "value10"),
 	}
-	fiveLogrusFields mlog.Fields
+	oneMlogField   = tenMlogFields[:1]
+	fiveMlogFields = tenMlogFields[:5]
 )
 
-func init() {
-	// Helper to create a 5-field map from the 10-field map.
-	fiveLogrusFields = make(mlog.Fields, 5)
-	i := 0
-	for k, v := range tenLogrusFields {
-		if i >= 5 {
-			break
-		}
-		fiveLogrusFields[k] = v
-		i++
-	}
-}
-
-// setupBenchmarkLogger creates a logger that writes to io.Discard,
+// setupBenchmarkLogger creates a logger that writes to a null device,
 // effectively isolating the benchmark to the logger's processing overhead by eliminating I/O.
 func setupBenchmarkLogger(b *testing.B) *mlog.Logger {
 	b.Helper()
-	// mlog is well-designed for benchmarking, as disabling both Stdout and Filepath
-	// in the config will automatically set the output to io.Discard.
+	// Redirecting output to /dev/null is a standard technique for benchmarking I/O components
+	// without measuring the actual I/O performance. This is necessary because mlog does not
+	// currently provide a public API to set the output writer directly to io.Discard.
 	cfg := mlog.Config{
 		Level:    mlog.DebugLevel,
+		Filepath: os.DevNull, // Write to the null device to discard output.
 		Stdout:   false,
-		Filepath: "",
 		Format:   "json",
 	}
 
 	logger := mlog.New(&cfg)
+	b.Cleanup(func() {
+		_ = logger.Close()
+	})
 	return logger
 }
 
@@ -61,7 +54,7 @@ func BenchmarkMlog_Simple(b *testing.B) {
 	b.ResetTimer()
 	b.RunParallel(func(pb *testing.PB) {
 		for pb.Next() {
-			logger.Info(ctx, "a simple message")
+			logger.Infow(ctx, "a simple message")
 		}
 	})
 }
@@ -76,38 +69,45 @@ func BenchmarkMlog_Sprintf(b *testing.B) {
 	})
 }
 
+func BenchmarkMlog_With1Field(b *testing.B) {
+	logger := setupBenchmarkLogger(b)
+	b.ResetTimer()
+	b.RunParallel(func(pb *testing.PB) {
+		for pb.Next() {
+			logger.Infow(ctx, "message with one field", oneMlogField...)
+		}
+	})
+}
+
+func BenchmarkMlog_With5Fields(b *testing.B) {
+	logger := setupBenchmarkLogger(b)
+	b.ResetTimer()
+	b.RunParallel(func(pb *testing.PB) {
+		for pb.Next() {
+			logger.Infow(ctx, "message with five fields", fiveMlogFields...)
+		}
+	})
+}
+
 func BenchmarkMlog_With10Fields(b *testing.B) {
 	logger := setupBenchmarkLogger(b)
 	b.ResetTimer()
 	b.RunParallel(func(pb *testing.PB) {
 		for pb.Next() {
-			// In mlog, WithFields creates a new logger entry on each call.
-			logger.WithFields(tenLogrusFields).Info(ctx, "message with ten fields")
+			logger.Infow(ctx, "message with ten fields", tenMlogFields...)
 		}
 	})
 }
 
 // BenchmarkMlog_WithLogger10Fields tests the performance of a logger that has been pre-configured
-// with contextual fields using the WithFields() method. This is a very common and important use case.
+// with contextual fields using the With() method. This is a very common and important use case.
 func BenchmarkMlog_WithLogger10Fields(b *testing.B) {
 	logger := setupBenchmarkLogger(b)
-	withLogger := logger.WithFields(tenLogrusFields)
+	withLogger := logger.With(tenMlogFields...)
 	b.ResetTimer()
 	b.RunParallel(func(pb *testing.PB) {
 		for pb.Next() {
-			withLogger.Info(ctx, "message from a contextual logger")
-		}
-	})
-}
-
-// BenchmarkMlog_With10FieldsInArgs tests the unique mlog feature of passing fields
-// directly as variadic arguments, which involves runtime reflection.
-func BenchmarkMlog_With10FieldsInArgs(b *testing.B) {
-	logger := setupBenchmarkLogger(b)
-	b.ResetTimer()
-	b.RunParallel(func(pb *testing.PB) {
-		for pb.Next() {
-			logger.Info(ctx, "message with ten fields", tenLogrusFields)
+			withLogger.Infow(ctx, "message from a contextual logger")
 		}
 	})
 }
