+++ version.go
@@ -2,7 +2,7 @@ package maltose
 
 const (
 	// VERSION is the current maltose version.
-	VERSION = "v0.0.0"
+	VERSION = "v0.1.4"
 	// COMPONENT is the component name.
-	COMPONENT = "comp"
+	COMPONENT = "component"
 )
